Ignite Training Session 2/requirements.txt
boto3
duckduckgo_search
mcp[cli]
nova-act
opensearch-py
pandas
retrying
strands-agents 
strands-agents-tools[mem0_memory]
streamlit
tavily-python
tqdm
uv
strands-agents[litellm]